+++ src/ua/monitored_item_id.rs
@@ -14,14 +14,10 @@ impl MonitoredItemId {
         Self(id)
     }
 
+    /// Gets the generic [`ua::IntegerId`].
     #[must_use]
-    pub(crate) const fn as_id(self) -> ua::IntegerId {
+    pub const fn as_id(self) -> ua::IntegerId {
         self.0
-    }
-
-    #[must_use]
-    pub const fn is_valid(self) -> bool {
-        self.0.is_valid()
     }
 }